Reuploading existing project to new repository
This group-project is a booking website for meetings and contract creation. I've chosen this project to showcase my proficiency in Vue.js, JavaScript/TypeScript, Behaviour Driven Design, Google Firebase, and Vuetify.I believe it demonstrates my capability to deliver effective solutions and collaborate within a team setting, aligning well with the requirements of the internship positions I'm applying for.